Fixes #263
